requirements.txt
PyBluez
pylint
PyYAML
